+++ common/src/main/java/com/evandev/reliable_backpacks/common/events/EntityInteractionEvents.java
@@ -1,12 +1,11 @@
 package com.evandev.reliable_backpacks.common.events;
 
 import com.evandev.reliable_backpacks.common.items.BackpackItemContainer;
-import com.evandev.reliable_backpacks.registry.BPItems;
+import com.evandev.reliable_backpacks.platform.Services;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.SimpleMenuProvider;
 import net.minecraft.world.entity.Entity;
-import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.ShulkerBoxMenu;
@@ -17,10 +16,9 @@ public class EntityInteractionEvents {
 
     public static InteractionResult onEntityInteract(Player player, Entity targetEntity) {
         LivingEntity target = targetEntity instanceof LivingEntity ? (LivingEntity) targetEntity : null;
-        ItemStack item = target != null ? target.getItemBySlot(EquipmentSlot.CHEST) : null;
+        ItemStack item = target != null ? Services.PLATFORM.getBackpack(target) : ItemStack.EMPTY;
 
-        if (target != null && item.is(BPItems.BACKPACK) && isBehind(player, target)) {
-
+        if (target != null && !item.isEmpty() && isBehind(player, target)) {
             if (!player.level().isClientSide()) {
                 BackpackItemContainer container = new BackpackItemContainer(target, player);
                 player.openMenu(new SimpleMenuProvider((a, b, c) -> new ShulkerBoxMenu(a, player.getInventory(), container), Component.translatable("container.backpack")));