+++ common/src/main/java/com/evandev/reliable_backpacks/mixin/SlotMixin.java
@@ -3,7 +3,6 @@ package com.evandev.reliable_backpacks.mixin;
 import com.evandev.reliable_backpacks.registry.BPItems;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.world.entity.player.Inventory;
-import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.Slot;
 import net.minecraft.world.item.ItemStack;
 import org.spongepowered.asm.mixin.Mixin;
@@ -18,22 +17,6 @@ public abstract class SlotMixin {
 
     @Shadow
     public abstract ItemStack getItem();
-
-    @Inject(
-            method = "mayPickup",
-            at = @At("HEAD"),
-            cancellable = true
-    )
-    public void mayPickup(Player player, CallbackInfoReturnable<Boolean> cir) {
-        Slot thisSlot = (Slot) (Object) this;
-
-        if (thisSlot.container instanceof Inventory && thisSlot.getContainerSlot() == 38) {
-            ItemStack item = this.getItem();
-            if (isNonEmptyBackpack(item)) {
-                cir.setReturnValue(false);
-            }
-        }
-    }
 
     @Inject(
             method = "mayPlace",