+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - UNRELEASED
+## [1.1.0] - 2026-03-13
+
+### Added
+
+- Added Curios integration (Forge) and Trinkets integration (Fabric).
+    - Backpacks can now be placed in the back slot when either mod is installed.
+
+## [1.0.0] - 2026-03-12
 
 - Initial release.